ModalCreatePost component created to pull off logic from modal component.
